Updated 'check-deps' to allow for more flexible dependency directories.
'check-deps' assumed that directories pointed to by the top level .rept-deps
file would match those pointed to by the .rept-deps files in the dependencies
(i.e. they woudl end up pointing to the same absolute path), but that is not
necessarily the case, so we account for the fact that all repos are checked out
by the top level, so we have to always use those paths to point to the
dependencies.